removing .env file
