--- .env
@@ -1,2 +1,0 @@
-USERNAME = 
-PASSWORD =
