+++ package.json
@@ -5,9 +5,6 @@
   "main": "index.js",
   "scripts": {
     "test": "npx playwright test",
-    "task-1": "npx playwright test ./src/tests/task1.spec.js",
-    "task-2": "npx playwright test ./src/tests/task2.spec.js",
-    "task-3": "npx playwright test ./src/tests/task3.spec.js",
     "test-ui": "npx playwright test --ui",
     "lint": "npx eslint . --color --fix --quiet"
   },
